+++ src/utils/hmac.cpp
@@ -7,9 +7,9 @@
 #include <string>
 
 bool verify_sign(const std::string& sign, const std::string& data) {
-  return sign == hmac<SHA256>(data, HMAC_SECRET);
+  return sign == hmac<SHA256>(data, ENCRYPTION_KEY);
 }
 
 std::string generate_sign(const std::string& data) {
-  return hmac<SHA256>(data, HMAC_SECRET);
+  return hmac<SHA256>(data, ENCRYPTION_KEY);
 }
